services/ingestion_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
elasticsearch>=8.15.0
openai>=1.40.0
langchain-text-splitters>=0.3.0
tiktoken>=0.7.0
PyYAML>=6.0.2
requests>=2.32.0
httpx>=0.27.0
boto3>=1.35.0
pdfplumber>=0.11.4
PyMuPDF>=1.24.0
pdf2image>=1.17.0
pytesseract>=0.3.10
mistralai>=1.2.0
python-docx>=1.1.2
odfpy>=1.4.1
olefile>=0.47
openpyxl>=3.1.5
numpy>=2.1.0
networkx>=3.3
Pillow>=10.4.0
